refactor(theme): apply simplify review cleanups
- Extract applyThemeToDOM into its own React-free module so the test
  needs no runtime mock or testing-only export
- Move the scrollbar suppression next to the transition it cancels in
  scrollbar.scss, scoped to .rp-scrollbar instead of the universal bucket
- Move the appearance view-transition rule into SwitchAppearance's
  global.scss where its sibling rules live
- Fold the suppression toggle into a single guarded block and mark the
  forced style flush with void so it does not read as dead code
