+++ packages/core/src/theme/styles/base.css
@@ -35,25 +35,15 @@ html.dark {
 }
 
 /* Toggled by applyThemeToDOM for one frame while the theme class flips, so
-   every themed surface switches in sync instead of fading at its own pace. */
+   every themed surface switches in sync instead of fading at its own pace.
+   A static rule toggled via a root class (rather than an injected <style>)
+   stays compatible with strict CSP. `*` does not match -webkit- scrollbar
+   pseudo-elements; their suppression lives next to their transition in
+   scrollbar.scss. */
 html.rp-theme-switching *,
 html.rp-theme-switching *::before,
 html.rp-theme-switching *::after {
   transition: none !important;
-}
-
-/* Scrollbar thumbs transition their background and are not matched by `*`.
-   Kept as a separate rule: a browser that does not recognize the
-   -webkit- pseudo-element would drop every selector sharing the list. */
-html.rp-theme-switching *::-webkit-scrollbar-thumb {
-  transition: none !important;
-}
-
-/* Toggled by SwitchAppearance while the appearance view transition runs; the
-   'flip' and 'root' view-transition names conflict otherwise, see
-   https://github.com/web-infra-dev/rspress/pull/1272 */
-html.rp-appearance-transition .rspress-doc {
-  view-transition-name: none !important;
 }
 
 body {
